refactor(daemon): extract runner::tracking submodule (#2642)
Move record_task_completion and record_task_failure to
runner/tracking.rs. Clean up imports: BuiltinTask and TaskAction
are now cfg(test)-only in mod.rs since only tests reference them
through `use super::*`.
